feat!: remove support for old versions of node and eslint
This change removes support for versions of `ESLint` less than 8.57, sets supported node versions to the currently active or maintenance LTS versions.

This set of versions is consistent with the versions supported by `eslint-plugin-eslint-plugin`, with the exception of including support for eslint 8.57.  I only kept that in the support list, since the repo is still using .eslintrc configs.  As part of #272, I suggest we bump compatibility to only versions 9.0 and higher.